+++ K-Content/src/main/resources/static/css/content/new-admin-cntnt-main.css
@@ -53,3 +53,7 @@
         width: 33.33333%;
     }
 }
+.categories{
+    display: flex;
+    justify-content: space-around;
+}
